pyproject.toml
[project]
name = "pulsestream-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiokafka>=0.12.0",
    "alembic>=1.17.1",
    "fastapi>=0.119.0",
    "jose>=1.0.0",
    "kafka-python>=2.2.15",
    "langdetect>=1.0.9",
    "pandas>=2.3.3",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2>=2.9.11",
    "pydantic[email]>=2.12.2",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "snowflake-connector-python[pandas]>=4.0.0",
    "sqlmodel",
    "textblob>=0.19.0",
    "uvicorn>=0.37.0",
]

[tool.uv.workspace]
members = [
    "api/utils",
]
